--- src/src/repository/GenericDao.java
@@ -1,17 +1,0 @@
-package repository;
-
-import java.sql.SQLException;
-import java.util.List;
-
-public interface GenericDao<E> {
-
-    void salvar (E objeto);
-    void remover (E objeto) throws SQLException, ClassNotFoundException;
-
-    List<E> buscarTodos() throws  SQLException, ClassNotFoundException;
-
-    List<E> buscarPorNomes(String nome);
-
-
-
-}
